requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
pdfminer.six>=20221105
tiktoken>=0.5.1
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
loguru>=0.7.2
chromadb
sentence-transformers==2.2.2
torch==2.2.2+cpu  # CPU-only version
numpy<2.0  # For compatibility with older compiled modules like torch/faiss
huggingface-hub==0.10.1 # For compatibility with sentence-transformers==2.2.2

# For backend API
Flask>=3.0.0
Flask-Cors>=4.0.0
